requirements.txt
discord.py==2.3.2
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydle==0.9.4
python-dotenv==1.0.1
httpx==0.25.2
aiohttp==3.10.5
feedparser==6.0.11
yt-dlp==2024.10.22
python-telegram-bot==20.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
atproto==0.0.54
pysnmp==4.4.12
pyasn1<0.6.0
cryptography==42.0.5
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
slowapi==0.1.9


